requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
Pillow==10.0.0
SpeechRecognition==3.10.0
pyaudio==0.2.11
translate==3.6.1
googletrans==4.0.0-rc1
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
google-generativeai==0.3.0